# Conflicts:
#	Smile_Android/app/src/main/java/com/ssafy/smile/common/util/Constants.kt
#	Smile_Android/app/src/main/java/com/ssafy/smile/data/DataSourceInstances.kt
#	Smile_Android/app/src/main/java/com/ssafy/smile/data/RepositoryInstances.kt
#	Smile_Android/app/src/main/java/com/ssafy/smile/domain/repository/PhotographerRepository.kt
#	Smile_Android/app/src/main/java/com/ssafy/smile/presentation/view/home/HomeFragment.kt


Former-commit-id: f3ada94bd6992cb97cdf043012bb744a857ad077

+++ Smile_Android/app/src/main/java/com/ssafy/smile/data/remote/datasource/PortfolioRemoteDataSourceImpl.kt
@@ -18,12 +18,12 @@ class PortfolioRemoteDataSourceImpl(private val portfolioApiService: PortfolioAp
     override suspend fun uploadPost(images: MutableMap<String, RequestBody>) : Response<Any>{
         return portfolioApiService.uploadPost(images)
     }
-    override suspend fun uploadPost(latitude: Float, longitude: Float,
+    override suspend fun uploadPost(latitude: Double, longitude: Double,
                                     detailAddress: String, category: String, images : List<MultipartBody.Part>): Response<Any>{
         return portfolioApiService.uploadPost(latitude, longitude, detailAddress, category, images)
     }
 
-    override suspend fun uploadPost(latitude: Float, longitude: Float,
+    override suspend fun uploadPost(latitude: Double, longitude: Double,
                                     detailAddress: String, category: String, images : HashMap<String, List<MultipartBody.Part>>): Response<Any>{
         return portfolioApiService.uploadPost(latitude, longitude, detailAddress, category, images)
     }
